Déplace le code relatif aux tags dans un fichier dédié
