+++ zds/tutorialv2/urls/urls_contents.py
@@ -35,7 +35,8 @@ from zds.tutorialv2.views.contributors import (
     RemoveContributorFromContent,
     ContentOfContributors,
 )
-from zds.tutorialv2.views.editorialization import RemoveSuggestion, AddSuggestion, EditContentTags
+from zds.tutorialv2.views.editorialization import RemoveSuggestion, AddSuggestion
+from zds.tutorialv2.views.tags import EditTags
 
 from zds.tutorialv2.views.lists import TagsListView, ContentOfAuthor, ListContentReactions
 from zds.tutorialv2.views.alerts import SendContentAlert, SolveContentAlert
@@ -208,7 +209,7 @@ urlpatterns = (
         # Modify the license
         path("modifier-licence/<int:pk>/", EditContentLicense.as_view(), name="edit-license"),
         # Modify the tags
-        path("modifier-tags/<int:pk>/", EditContentTags.as_view(), name="edit-tags"),
+        path("modifier-tags/<int:pk>/", EditTags.as_view(), name="edit-tags"),
         # beta:
         path("activer-beta/<int:pk>/<slug:slug>/", ManageBetaContent.as_view(action="set"), name="set-beta"),
         path(
